backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.9,<3
pydantic-settings>=2.5
httpx>=0.27
python-dotenv>=1.0
PyJWT[crypto]>=2.9
cryptography>=43
google-genai>=0.3
groq>=0.18
supabase>=2.9
tenacity>=9
structlog>=24
